// packages/core/enums/exceptions.ts
export enum NodecordExceptionCode {
    CIRCULAR_IMPORT = 'CIRCULAR_IMPORT',
    INVALID_MODULE = 'INVALID_MODULE',
    INVALID_PROVIDER = 'INVALID_PROVIDER',
    INVALID_LISTENER = 'INVALID_LISTENER',
    INVALID_HANDLER = 'INVALID_HANDLER',
    INVALID_INTERCEPTOR = 'INVALID_INTERCEPTOR',
    PROVIDER_NOT_FOUND = 'PROVIDER_NOT_FOUND',
    DUPLICATE_INTERCEPTOR = 'DUPLICATE_INTERCEPTOR',
    INVALID_EXCEPTION_HANDLER = 'INVALID_EXCEPTION_HANDLER',
    MISSING_CONTRACT_METHOD = 'MISSING_CONTRACT_METHOD',
    INTERNAL_ERROR = 'INTERNAL_ERROR',
    MISSING_CLIENT_OPTIONS = 'MISSING_CLIENT_OPTIONS',
}